Uninstall curl before running debian-bootstrap
debian-bootstrap attempts to install cabal-install-1.16, which can
either work with curl, wget, or fetch, in that order of preference.
Unfortunately the hackage URLs have changed since 1.16 and with the
arguments used in cabal-install's bootstrap script, curl doesn't
follow 301 redirects.

The simplest solution is to uninstall curl, so that wget is used
instead.