Include expected test failures in build-plan.txt
Pinging @manny-fp. This change was necessary since the list of expected
failures will change over time, but old build plans should *not* depend
on updates to the list. This breaks the file format of build-plan.txt
slightly, though it is easy to fix old files. As an example:

-- BEGIN expected-failures
GLUT
HTF
HTTP
Octree
...
-- END expected-failures