Remove cabal-test-quickcheck, restore http-media
Packages tested with cabal-test-quickcheck cannot pass on all of the
required platforms: it requires that Cabal and cabal-install be on the
same side of version 1.19, but a build against GHC 7.8 will tend to use
Cabal-1.18 with cabal-install-1.20 or higher. As such, the library has
been removed. The http-media package, which used cabal-test-quickcheck,
is now expected to build its test suite and pass all tests.