add Win32-extras
This is a dependency of git-annex on Windows, and won't otherwise be included since deps are resolved for linux and not windows. Per https://github.com/commercialhaskell/stack/issues/1158

I put it under my name, although I don't maintain the package, because I can't speak for the authors of Win32-extras regarding putting it in stackage, and it's so far only included because git-annex uses it. Hopefully this is ok.