HDBC-mysql: fix package name capitalization
I noticed https://www.stackage.org/package/HDBC-mysql doesn't list this package in any snapshots, but I've had it in the build-constraints.yaml file for a long time. I'm guessing it is because the capitalization was wrong, but maybe someone with more knowledge of Stackage can confirm that, or tell me why it's not in any snapshot.