Add haskeline.
It builds fine with ghc-8.6.1, and it was in previous LTS's with no issues, and
so i'm not sure why it was removed in the ghc-8.6 nightlies.   Might be
something to do with being distributed with GHC, but not being a core package
from Stackage's perspective.  In either case, adding it explicitly seems like
the right thing to do long-term.