Reenable vault
`vault` was listed in `"GHC 8.8-related build failures"`, but that is
much too restrictive, since the library actually builds without issue
on GHC 8.8. The only part that doesn't build correctly is its
Haddocks, which is due to a Haddock bug (reported upstream at
haskell/haddock#1091). I've listed `vault` in
`expected-haddock-failures` as a result.