Add reflex, jsaddle, ghcjs-dom, and related packages
This adds Reflex to Stackage as well as some packages that are
commonly used with Reflex. The goal is to make something like
“reflex-todomvc”[1] work out-of-the-box with Stackage. The ghcjs and
jsaddle dependencies are needed for running Reflex on a development
machine, but do not themselves require the ghcjs compiler.

These packages are fairly stable, and should be compatible with other
Stackage dependencies. I am not the maintainer for any of these
packages, but can work in keeping these up-to-date.

[1]: https://github.com/reflex-frp/reflex-todomvc