Remove constraint on universum
Problem: universum-1.6.0 and newer doesn't support text-1.2.4.1
which is the latest version of text.
It was reported in https://github.com/commercialhaskell/stackage/issues/5796
Because of that `build-constraints.yaml` now constrains universum's
version to be less than 1.6.0.

Solution: now that universum-1.7.2 has been released and supports the
latest `text`, the constraint can be dropped entirely.