Add optparse-applicative 0.18 compatibility flags
While previous versions of `optparse-applicative` used the
`ansi-wl-pprint` library for formatting text, the `prettyprinter`
library is used instead from version `0.18.0.0`.  A
`optparse-applicative_ge_0_18` flag is used to set the correct
dependency constraints according to the version of
`optparse-applicative` used.

This commit sets the flag to `true` in the following packages:

* `bm`
* `horizontal-rule`
* `literatex`
* `phatsort`
* `redact`